requirements.txt
streamlit
langchain
langchain-google-genai
langchain-anthropic
langchain-qdrant
langgraph
qdrant-client
python-dotenv
pypdf
duckduckgo-search
galileo
